Fixed compile errors

+++ directmemory-cache/src/test/java/org/apache/directmemory/memory/RoundRobinAllocationPolicyTest.java
@@ -138,7 +138,7 @@ public class RoundRobinAllocationPolicyTest
         }
 
         @Override
-        public int getCapacity()
+        public long getCapacity()
         {
             return 0;
         }
